+++ t/test_3_30.t
@@ -314,9 +314,14 @@ else
   open( TMP, ">$tmp") or die "cannot create temp file";
   XML::Twig->parse( twig_roots => { e1 => sub { $_->flush( \*TMP) } }, twig_print_outside_roots => \*TMP, $doc);
   close TMP;
-  my $res= slurp( $tmp);
-  is( $res, $doc, "bug in flush with twig_print_outside_roots");
-  unlink $tmp;
+  if ( -f $tmp ) {
+      my $res= slurp( $tmp);
+      is( $res, $doc, "bug in flush with twig_print_outside_roots");
+      unlink $tmp;
+  } else {
+      skip( 1,
+            "problem with writing $tmp, likely linked to missing write permission on the current directory" );
+  }
 }
 
 { # test bug where #default appeared in attributes (RT #27617)
